cpd: to correct failover behavior of cpsv [#1765]
Problem:
In case failover multiple times, the cpnd is down for a moment so there is no cpnd opening specific checkpoint.
This lead to retention timer is trigger.
When cpnd is up again but has different pid so retention timer is not stoped.
Repica is deleted at retention while its information still be in ckpt database.
That cause problem.

Fix:
- Stop timer of removed node.
- Update data in patricia trees (for retention value consistence).