amfd: Add iteration to failover all absent assignments [#2416]
Problem: As described in previous patch: "amfd: Make creation
and deletion of assignment object as IMM synced call [#2416]"

The previous patch makes creation and deletion assignment object
as IMM synced call at the first try. It acts as the prevention of
inconsistency between AMFND and IMM. However, if the first IMM
call fails, a chance of same problem can happen.

This patch continues the approach of absent assignments that is
using node_fail() Sg Fsm to perform failover. This node_fail()
now also is used as a clean up method that iterates and remove
all absent invalid assignments as well as failover valid assignments.